Port ValueError to new error mechanism as well
I am not entirely happy with how the reporting of the value path looks,
but for now it's good enough. Let's first make a pretty-printer for RCL
values so we can begin including values in error messages.
