+++ src/lib.rs
@@ -14,7 +14,6 @@ pub mod abstraction;
 pub mod ast;
 pub mod cst;
 pub mod error;
-pub mod error_old;
 pub mod eval;
 pub mod fmt;
 pub mod highlight;
